requirements.txt
authlib ~= 1.0
django ~= 4.0
python-dotenv ~= 0.19
requests ~= 2.27
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
django-auth0==0.0.3
django-google-sso==7.0.0
djangorestframework==3.15.2
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
httplib2==0.22.0
idna==3.8
loguru==0.7.2
oauthlib==3.2.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyparsing==3.1.4
python-dotenv==0.19.2
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
